requirements.txt
fastapi[standard]>=0.110.0
sqlalchemy>=2.0.0,<3.0.0
aiomysql>=0.2.0
cryptography>=42.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
minio>=7.2.0
elasticsearch>=8.13.4
uvicorn[standard]>=0.27.0
pytest>=8.0.0
httpx>=0.26.0
aiohttp>=3.9.0
ldap3>=2.9.1
redis>=5.0.0
pyjwt>=2.8.0
python-multipart>=0.0.9
